+++ src/api/favorites/favorites.module.ts
@@ -3,9 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Module, forwardRef } from '@nestjs/common';
 import { FavoritesService } from './favorites.service';
 import { FavoritesController } from './favorites.controller';
-import { ArtistModule } from '../artist/artist.module';
-import { AlbumModule } from '../album/album.module';
-import { TrackModule } from '../track/track.module';
+import { ArtistModule, AlbumModule, TrackModule } from '..';
 
 @Module({
   imports: [
