+++ src/api/index.ts
@@ -2,9 +2,6 @@ import { Artist } from './artist/entity/artist.entity';
 import { Album } from './album/entity/album.entity';
 import { Track } from './track/entity/track.entity';
 import { Favorites } from './favorites/entity/favorites.entity';
-// import { ArtistService } from './artist/artist.service';
-// import { AlbumService } from './album/album.service';
-// import { TrackService } from './track/track.service';
 
 export const enitities = [Artist, Album, Track, Favorites];
 export type Entity = Artist | Album | Track;
@@ -14,12 +11,16 @@ export enum FavEntity {
   Track = 'tracks',
 }
 export { Artist, Album, Track, Favorites };
+export { User } from './user/entity/user.entity';
 
 export { IArtist } from './artist/interfaces/artist.interface';
 export { IAlbum } from './album/interfaces/album.interface';
 export { ITrack } from './track/interfaces/track.interface';
 export { IFavorites } from './favorites/interfaces/favorites.interface';
+export { IUser } from './user/interfaces/user.interface';
 
-// export type EntityService = ArtistService | AlbumService | TrackService;
-// export { ArtistService, AlbumService, TrackService };
-// export { FavoritesService } from './favorites/favorites.service';
+export { UserModule } from './user/user.module';
+export { ArtistModule } from './artist/artist.module';
+export { AlbumModule } from './album/album.module';
+export { TrackModule } from './track/track.module';
+export { FavoritesModule } from './favorites/favorites.module';
