+++ src/app.module.ts
@@ -3,11 +3,13 @@ import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { ArtistModule } from './api/artist/artist.module';
-import { AlbumModule } from './api/album/album.module';
-import { TrackModule } from './api/track/track.module';
-import { FavoritesModule } from './api/favorites/favorites.module';
-import { UserModule } from './api/user/user.module';
+import {
+  UserModule,
+  ArtistModule,
+  AlbumModule,
+  TrackModule,
+  FavoritesModule,
+} from './api';
 import { configService } from './config/ormconfig';
 
 @Module({
