+++ src/config/ormconfig.ts
@@ -1,10 +1,6 @@
-import { Favorites } from './../api/favorites/entity/favorites.entity';
 import * as dotenv from 'dotenv';
 import { DataSourceOptions, DataSource } from 'typeorm';
-import { User } from '../api/user/entity/user.entity';
-import { Artist } from '../api/artist/entity/artist.entity';
-import { Track } from './../api/track/entity/track.entity';
-import { Album } from './../api/album/entity/album.entity';
+import { User, Artist, Album, Track, Favorites } from '../api';
 
 dotenv.config();
