+++ src/interceptors/index.ts
@@ -1,0 +1,2 @@
+export { HeadersInterceptor } from './headers.interceptor';
+export { NotFoundInterceptor } from './not-found.interceptor';
